Rework file system and rename it
