+++ src/adapter/fileSystem/FileSystemService.ts
@@ -1,4 +1,5 @@
 export interface FileSystemService {
   createDirectory(directoryPath: string): Promise<void>;
   readFile(filePath: string): Promise<string>;
+  writeFile(filePath: string, content: Uint8Array): Promise<void>;
 }
